+++ test/videotracker.spec.js
@@ -1,4 +1,4 @@
-import Tracker from '../src/videotracker'
+import VideoTracker from '../src/videotracker'
 import Log from '../src/log'
 import chai from 'chai'
 import sinon from 'sinon'
@@ -6,7 +6,7 @@ import sinon from 'sinon'
 const expect = chai.expect
 const assert = chai.assert
 
-describe('VideoTracker', () => {
+describe('VideoVideoTracker', () => {
   let tracker
 
   // Mute console
@@ -20,7 +20,7 @@ describe('VideoTracker', () => {
 
   describe('settings', () => {
     it('should set options', () => {
-      tracker = new Tracker(1, { isAd: true, tag: 2 })
+      tracker = new VideoTracker(1, { isAd: true, tag: 2 })
       expect(tracker.isAd()).to.be.true
       expect(tracker.player).to.equal(1)
       expect(tracker.tag).to.equal(2)
@@ -32,8 +32,8 @@ describe('VideoTracker', () => {
     })
 
     it('should send custom data', (done) => {
-      tracker = new Tracker(null, { customData: { a: 1 } })
-      tracker.on(Tracker.Events.PLAYER_INIT, (e) => {
+      tracker = new VideoTracker(null, { customData: { a: 1 } })
+      tracker.on(VideoTracker.Events.PLAYER_INIT, (e) => {
         expect(e.data.a).to.equal(1)
         done()
       })
@@ -41,8 +41,8 @@ describe('VideoTracker', () => {
     })
 
     it('should set adsTracker', (done) => {
-      tracker = new Tracker(null, { adsTracker: new Tracker() })
-      tracker.adsTracker.on('AD_' + Tracker.Events.REQUEST, () => {
+      tracker = new VideoTracker(null, { adsTracker: new VideoTracker() })
+      tracker.adsTracker.on('AD_' + VideoTracker.Events.REQUEST, () => {
         tracker.disposeAdsTracker()
         done()
       })
@@ -51,7 +51,7 @@ describe('VideoTracker', () => {
 
     it('should calculate webkitbitrate', () => {
       let tag = { webkitVideoDecodedByteCount: 1000 }
-      tracker = new Tracker(tag)
+      tracker = new VideoTracker(tag)
       expect(tracker.getWebkitBitrate()).to.be.null
 
       tag.webkitVideoDecodedByteCount += 3000
@@ -61,53 +61,53 @@ describe('VideoTracker', () => {
 
   describe('EventFiring', () => {
     beforeEach(() => {
-      tracker = new Tracker()
+      tracker = new VideoTracker()
     })
 
     it('player init', (done) => {
-      tracker.on(Tracker.Events.PLAYER_INIT, () => done())
+      tracker.on(VideoTracker.Events.PLAYER_INIT, () => done())
       tracker.sendPlayerInit()
     })
 
     it('player ready', (done) => {
-      tracker.on(Tracker.Events.PLAYER_READY, () => done())
+      tracker.on(VideoTracker.Events.PLAYER_READY, () => done())
       tracker.sendPlayerInit()
       tracker.sendPlayerReady()
     })
 
     it('download', (done) => {
-      tracker.on(Tracker.Events.DOWNLOAD, () => done())
+      tracker.on(VideoTracker.Events.DOWNLOAD, () => done())
       tracker.sendDownload()
     })
 
     // Video
     it('request', (done) => {
-      tracker.on('CONTENT_' + Tracker.Events.REQUEST, () => done())
+      tracker.on('CONTENT_' + VideoTracker.Events.REQUEST, () => done())
       tracker.sendRequest()
     })
 
     it('start', (done) => {
-      tracker.on('CONTENT_' + Tracker.Events.START, () => done())
+      tracker.on('CONTENT_' + VideoTracker.Events.START, () => done())
       tracker.sendRequest()
       tracker.sendStart()
     })
 
     it('end', (done) => {
-      tracker.on('CONTENT_' + Tracker.Events.END, () => done())
+      tracker.on('CONTENT_' + VideoTracker.Events.END, () => done())
       tracker.sendRequest()
       tracker.sendStart()
       tracker.sendEnd()
     })
 
     it('pause', (done) => {
-      tracker.on('CONTENT_' + Tracker.Events.PAUSE, () => done())
+      tracker.on('CONTENT_' + VideoTracker.Events.PAUSE, () => done())
       tracker.sendRequest()
       tracker.sendStart()
       tracker.sendPause()
     })
 
     it('resume', (done) => {
-      tracker.on('CONTENT_' + Tracker.Events.RESUME, () => done())
+      tracker.on('CONTENT_' + VideoTracker.Events.RESUME, () => done())
       tracker.sendRequest()
       tracker.sendStart()
       tracker.sendPause()
@@ -115,14 +115,14 @@ describe('VideoTracker', () => {
     })
 
     it('seek start', (done) => {
-      tracker.on('CONTENT_' + Tracker.Events.SEEK_START, () => done())
+      tracker.on('CONTENT_' + VideoTracker.Events.SEEK_START, () => done())
       tracker.sendRequest()
       tracker.sendStart()
       tracker.sendSeekStart()
     })
 
     it('seek end', (done) => {
-      tracker.on('CONTENT_' + Tracker.Events.SEEK_END, () => done())
+      tracker.on('CONTENT_' + VideoTracker.Events.SEEK_END, () => done())
       tracker.sendRequest()
       tracker.sendStart()
       tracker.sendSeekStart()
@@ -130,14 +130,14 @@ describe('VideoTracker', () => {
     })
 
     it('buffer start', (done) => {
-      tracker.on('CONTENT_' + Tracker.Events.BUFFER_START, () => done())
+      tracker.on('CONTENT_' + VideoTracker.Events.BUFFER_START, () => done())
       tracker.sendRequest()
       tracker.sendStart()
       tracker.sendBufferStart()
     })
 
     it('buffer end', (done) => {
-      tracker.on('CONTENT_' + Tracker.Events.BUFFER_END, () => done())
+      tracker.on('CONTENT_' + VideoTracker.Events.BUFFER_END, () => done())
       tracker.sendRequest()
       tracker.sendStart()
       tracker.sendBufferStart()
@@ -145,44 +145,44 @@ describe('VideoTracker', () => {
     })
 
     it('heartbeat', (done) => {
-      tracker.on('CONTENT_' + Tracker.Events.HEARTBEAT, () => done())
+      tracker.on('CONTENT_' + VideoTracker.Events.HEARTBEAT, () => done())
       tracker.sendRequest()
       tracker.sendHeartbeat()
     })
 
     it('rendition change', (done) => {
-      tracker.on('CONTENT_' + Tracker.Events.RENDITION_CHANGE, () => done())
+      tracker.on('CONTENT_' + VideoTracker.Events.RENDITION_CHANGE, () => done())
       tracker.sendRenditionChanged()
     })
 
     it('error', (done) => {
-      tracker.on('CONTENT_' + Tracker.Events.ERROR, () => done())
+      tracker.on('CONTENT_' + VideoTracker.Events.ERROR, () => done())
       tracker.sendError()
     })
 
     // Ads only
     it('ad break start', (done) => {
       tracker.setIsAd(true)
-      tracker.on(Tracker.Events.AD_BREAK_START, () => done())
+      tracker.on(VideoTracker.Events.AD_BREAK_START, () => done())
       tracker.sendAdBreakStart()
     })
 
     it('ad break end', (done) => {
       tracker.setIsAd(true)
-      tracker.on(Tracker.Events.AD_BREAK_END, () => done())
+      tracker.on(VideoTracker.Events.AD_BREAK_END, () => done())
       tracker.sendAdBreakStart()
       tracker.sendAdBreakEnd()
     })
 
     it('ad quartile', (done) => {
       tracker.setIsAd(true)
-      tracker.on(Tracker.Events.AD_QUARTILE, () => done())
+      tracker.on(VideoTracker.Events.AD_QUARTILE, () => done())
       tracker.sendAdQuartile()
     })
 
     it('ad click', (done) => {
       tracker.setIsAd(true)
-      tracker.on(Tracker.Events.AD_CLICK, () => done())
+      tracker.on(VideoTracker.Events.AD_CLICK, () => done())
       tracker.sendAdClick()
     })
   })